:hammer: Explode ern-orchestrator utils
