+++ ern-core/src/getDefaultMavenLocalDirectory.ts
@@ -1,0 +1,7 @@
+import os from 'os'
+import path from 'path'
+
+export function getDefaultMavenLocalDirectory() {
+  const pathToRepository = path.join(os.homedir(), '.m2', 'repository')
+  return `file://${pathToRepository}`
+}
